requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.10
dataclasses==0.6
datasets==2.18.0
debugpy==1.8.0
decorator==5.1.1
deepchem==2.8.0
descriptastorus==2.6.1
dgl==2.1.0
dill==0.3.8
docker-pycreds==0.4.0
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
flake8==7.0.0
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2023.12.2
fuzzywuzzy==0.18.0
gdown==4.7.1
gensim==4.3.2
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.62.1
huggingface-hub==0.20.2
-e git+https://github.com/gmum/huggingmolecules.git@85d1b9094dc02f9b776a9c61d73ae9099d66cf29#egg=huggingmolecules&subdirectory=src
idna==3.6
iniconfig==2.0.0
ipykernel==6.28.0
ipython==8.20.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
lightning-utilities==0.10.1
littleutils==0.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
-e git+https://github.com/BenevolentAI/MolBERT.git@b410cb6c98133b68789f368cc90adff6cb04418a#egg=molbert
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy==1.9.0
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
numpy==1.24.4
ogb==1.3.6
outdated==0.2.2
packaging==23.2
pandarallel==1.6.4
pandas==2.1.4
pandas-flavor==0.6.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
pluggy==1.4.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pyarrow-hotfix==0.6
pycodestyle==2.11.1
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
pyflakes==3.2.0
pyg-lib==0.3.1+pt21cpu
Pygments==2.17.2
pyparsing==3.1.1
PySocks==1.7.1
PyTDC==0.4.1
pytest==8.1.1
python-dateutil==2.8.2
pytorch-lightning==2.2.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
rdkit==2023.9.2
rdkit-pypi==2022.9.5
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.1
selfies==2.1.1
sentencepiece==0.2.0
sentry-sdk==1.44.1
seqeval==1.2.2
setproctitle==1.3.3
simpletransformers==0.63.9
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
soupsieve==2.5
stack-data==0.6.3
streamlit==1.33.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.2.0
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.1.2
torch_geometric==2.4.0
torchdata==0.7.1
torchmetrics==1.3.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.39.3
triton==2.1.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
wandb==0.13.10
watchdog==4.0.0
wcwidth==0.2.13
Werkzeug==3.0.1
xarray==2023.12.0
xxhash==3.4.1
yarl==1.9.4
